+++ CHANGELOG.md
@@ -15,6 +15,10 @@ version of DSCResource.Common
 
 - Added CMHierarchySetting resource
 
+- CMDistributionGroup added properties to configure collection membership in Distribution Point Group
+
+- Pin old version of ChangeLogManagement module to ensure module build.
+
 ## [3.0.0] - 2022-01-03
 
 ### Added